Create ML pipeline: processing, featurization, training, eval
Processed datasets are saved in the folder data/processed, and
serialized objects (encoders and models) in artifacts/. Note that in the
pipeline we take care of fitting the feature and label encoders on the
training dataset only.

The task for the ML pipeline is to classify the sex penguins using their
species as categorical feature (one-hot encoded) and their bill length,
bill depth, flipper length and body mass as numerical features.